pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flake8>=7.1.1",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
]
